Update: generate Mac bundle on install rather than packaging
This is a major overhaul of how OpenTTD is packaged for macOS. Rather
than using a disk image, it's now a Zip archive. The primary
motiviation for this was:

1) Disk images are annoying.
2) Technically, creating the bundle _is_ part of the build.

I noticed that OpenTTD set the working directory as part of its app
launch script. It seemed much simpler to me to simply do this in the
C++ code.

I also added an option for disabling the app bundle "fixup". This
should be useful for developers or packagers such as MacPorts -- which
was how I started looking at this.